all: Add IN_865_867 band definition
